# phase2-ai-enhanced/backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EquipShare API",
    "version": "1.0.0",
    "description": "API documentation for EquipShare (Phase 2)."
  },
  "servers": [{ "url": "http://localhost:5001/api" }],
  "paths": {
...
    "/requests": {
      "get": {
        "summary": "List requests",
        "tags": ["Requests"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Create request",
        "tags": ["Requests"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/requests/overdue": {
      "get": {
        "summary": "List overdue requests (admin/staff)",
        "tags": ["Requests"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/requests/{id}": {
      "get": {
        "summary": "Get request",
        "tags": ["Requests"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete request",
        "tags": ["Requests"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } }
      }
    },
    "/requests/{id}/status": {
      "put": {
        "summary": "Update request status (admin/staff)",
        "tags": ["Requests"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}